first version completed

+++ design/models.py
@@ -36,7 +36,7 @@ class parts(models.Model):
     sequence         = models.TextField(null=True)
     sequence_length  = models.IntegerField(null=True)
     part_url         = models.CharField(max_length=255, null=True)
-    score            = models.IntegerField(null=True)
+    score            = models.FloatField(null=True)
 
     def __unicode__(self):
         return self.part_name
